--- a/src/NServiceBus.Core.Tests/packages.config
+++ b/src/NServiceBus.Core.Tests/packages.config
@@ -3,6 +3,9 @@
   <package id="ApiApprover" version="3.0.1" targetFramework="net45" />
   <package id="ApprovalTests" version="3.0.11" targetFramework="net452" />
   <package id="ApprovalUtilities" version="3.0.11" targetFramework="net452" />
+  <package id="BenchmarkDotNet" version="0.9.7" targetFramework="net452" />
+  <package id="BenchmarkDotNet.Diagnostics.Windows" version="0.9.7" targetFramework="net452" />
+  <package id="Microsoft.Diagnostics.Tracing.TraceEvent" version="1.0.41" targetFramework="net452" />
   <package id="Mono.Cecil" version="0.9.6.1" targetFramework="net452" />
   <package id="NuDoq" version="1.2.5" targetFramework="net45" />
   <package id="NUnit" version="3.2.1" targetFramework="net452" />