--- a/src/NServiceBus.Msmq.AcceptanceTests/packages.config
+++ b/src/NServiceBus.Msmq.AcceptanceTests/packages.config
@@ -1,4 +1,5 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
+  <package id="BenchmarkDotNet" version="0.9.7" targetFramework="net46" />
   <package id="NUnit" version="3.2.1" targetFramework="net46" />
 </packages>