This PR spikes a different approach of deserializing headers. Currently it's MSMQ only, but this serializerless approach can be provided for others as well. This spike is based on:

1. No more decoding of the whole string with UTF8.
1. `XmlSerializer` replaced with a custom code, using internalized headers keys. It uses comparison of UTF8 encoded `ArraySegment<byte>` keys rather than decoding then over and over again. If UTF8 representation is equal, then an already allocated string is used. This decreases the execution time (no decoding) as well as # of allocated bytes.

See the results below:
```
BenchmarkDotNet=v0.9.7.0
OS=Microsoft Windows NT 10.0.10586.0
Processor=Intel(R) Core(TM) i7-6700HQ CPU 2.60GHz, ProcessorCount=8
Frequency=2531250 ticks, Resolution=395.0617 ns, Timer=TSC
HostCLR=MS.NET 4.0.30319.42000, Arch=64-bit RELEASE [RyuJIT]
JitModules=clrjit-v4.6.1080.0

Type=HeadersPerformanceTests  Mode=Throughput  
```

  Method |     Median |    StdDev |  Gen 0 | Gen 1 | Gen 2 | Bytes Allocated/Op |
-------- |----------- |---------- |------- |------ |------ |------------------- |
 Current | 26.4583 us | 1.9512 us | 914.00 |     - |     - |          10,540.56 |
    Fast |  9.6449 us | 0.4290 us | 166.00 |     - |     - |           1,620.50 |