@bording 2016-06-30
Unless this would be a step to a greater perf benefit to come, I'm not sure this would be worth doing. As long as the `XmlReader` is required, I don't see there being a lot of parsing improvement happening.

Around the area of headers, I wonder if there is more opportunity for allocation saving vs speed increase. Instead of needing to create a dictionary per message, what if we used a less heavy collection? Or, maybe pooling instances in some way and reusing them?