@bording 2016-06-30
Also, can you add the memory diagnoser to BenchmarkDotNet so we can see the allocation differences here?