@danielmarbach 2016-06-30
I think it is not enough to compare just the perf. Much more important is the heapload of garbage that is created and then the pressure we get due to Garbage collection. The more allocations we get the more we will suffer from GC pauses especially with higher concurrency values.