@Scooletz 2016-06-30
@danielmarbach Yes, and this allocated at least one `string` less. As always, I've been thinking about it and going with Kestrel approach, so encoding to UTF8 bytes key values, imo is the way to go.