@Scooletz 2016-07-01
I've updated the description. Without `XmlSerializer` at all, it's 2.5 times faster. No alloc tests still. Will add internalization of the keys.