@Scooletz 2016-07-01
Adding string memoization improved performance again. I'll add the memory benchmarks.

  Method |     Median |    StdDev |
-------- |----------- |---------- |
 Current | 24.9144 us | 3.5558 us |
    Fast |  8.7676 us | 0.1111 us |