@Scooletz 2016-07-01
With memory benchmarks. Just wanted to point out that values are still of type `string` and that we're returning a `Dictionary<string,string>` We should not expect 0 in there ;-)

  Method |     Median |    StdDev |  Gen 0 | Gen 1 | Gen 2 | Bytes Allocated/Op |
-------- |----------- |---------- |------- |------ |------ |------------------- |
 Current | 25.3901 us | 3.3454 us | 968.00 |     - |     - |          10,416.34 |
    Fast |  9.2381 us | 0.4564 us | 163.78 |     - |     - |           1,495.37 |