@danielmarbach 2016-07-01
What are the implications if making NSB unsafe?

On Friday, July 1, 2016, Szymon Kulec <notifications@github.com> wrote:

> Ping @danielmarbach <https://github.com/danielmarbach> @bording
> <https://github.com/bording>
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/Particular/NServiceBus/pull/3895#issuecomment-229896326>,
> or mute the thread
> <https://github.com/notifications/unsubscribe/AAKosuChuHnc67UvbE8o33dAi31gEZZyks5qRNgmgaJpZM4JCVga>
> .
>