@bording 2016-07-01
@Scooletz Very nice! The additional changes make a big difference! Seems like another possible step would be to look into some sort of object pooling instead of creating new `Dictionary<string,string>` all the time. That would be a larger change that would impact more areas, though.

> What are the implications if making NSB unsafe?

AFAIK, `<AllowUnsafeBlocks>true</AllowUnsafeBlocks>` shouldn't have any impact other than allowing `unsafe` usage.