@Scooletz 2016-07-04
Unsafe enables to use `unsafe` and pointers. Actually, I was amazed that NSB haven't had it configured yet :P