@SimonCropp 2016-07-05
Given .net core is on the horizon is this approach future proof?