@SimonCropp 2016-07-05
Is this code covered by some tests that verify behaviour under corrupted header conditions?