@Scooletz 2016-07-06
@SimonCropp 

> Given .net core is on the horizon is this approach future proof?

It's mostly about `unsafe`, no other dependencies. It's `.net core ready`

> Is this code covered by some tests that verify behaviour under corrupted header conditions?

I've provided one test only for the equivalent output to a regular method. I think that with current approach there may be cases that corrupted headers will be interpreted as valid. For instance value that contains `>` will be accepted as the next character that is searched for is `<`. Let review the code one more time and come with either more tests and/or more meaningful description of the possible scenarios.