@Scooletz 2016-07-06
@SimonCropp 

Let me list out different cases where the parser will fail, where by a failure I mean the false positively parsed xml or a false negatively thrown exception. All the cases will modify the properly formatted:
```
<HeaderInfo>
   <Key>NServiceBus.MessageIntent</Key>
   <Value>Subscribe</Value>
</HeaderInfo>
```
Frankly speaking, I don't know if these test constitute any test cases that are valid. I don't know the cases where the headers could be corrupted either.

#### < or > in the key, value (false positive)
The example below, containing an additional `>` will be parsed properly by the fast deserializer to a value `Subscribe>`. 
`XmlSerializer` would throw an exception.
```
<HeaderInfo>
   <Key>NServiceBus.MessageIntent</Key>
   <Value>Subscribe></Value>
</HeaderInfo>
```
#### additonal spaces (false negative)
Spaces after a tag name will break the fast deserialization when `XmlSerializer` would parse it correctly.
```
<HeaderInfo>
   <Key
>NServiceBus.MessageIntent</Key        >
   <Value

>Subscribe</Value                  >
</HeaderInfo>
```