@Scooletz 2016-07-11
@danielmarbach What should we do with this PR after providing my failure scenarios? Should we throw it away?