@danielmarbach 2016-07-11
To be honest I'm lacking experience what could go wrong for headers. Maybe @andreasohlund @johnsimons or @indualagarsamy can shed some light if it is "safe" to go ahead.