@johnsimons 2016-07-11
@Scooletz @danielmarbach I have read the discussion and looked at the code and it seems some serious performance boost so well done @Scooletz, but ....

The code is more complex, in the end we are still writing/reading xml.
IMO I would not implement this, I would instead change the serialization format to something else that provides the same benefits, and make it a opt-in option.