@Scooletz 2016-07-12
@johnsimons It's more complex because it supports `xml` ;-)

If providing a new opt-it serialization is an option, I'd say that we could do it this way. It'd be easy to check if it's xml and provide a different header for this new serialization.

Feel free to drop this PR.