@Scooletz 2016-07-14
I'm closing this PR as a result of the points mentioned in https://github.com/Particular/NServiceBus/pull/3895#issuecomment-231889788 This does not mean that we should avoid Kestrel trickery at all ;-)