@danielmarbach 2016-07-14
@Scooletz I don't think complexity was ever an issue that we didn't do something. A lot of the things in core are quite complex. As long as we can justify why it has that complexity and the code is still maintainable I see no reason not to do it. Let's hash it out on a call.